backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
passlib
bcrypt
pyjwt